// src/models/api/validate-address/validate_address_request_body.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type AddressToValidate = PartialAddress;
export type PostalCode = string;
export type CountryCode = string;
export type AddressResidentialIndicator = 'unknown' | 'yes' | 'no';
export type ValidateAddressRequestBody = AddressToValidate[];

export interface PartialAddress {
  name?: string;
  phone?: string;
  company_name?: string;
  address_line1?: string;
  address_line2?: string;
  address_line3?: string;
  city_locality?: string;
  state_province?: string;
  postal_code?: PostalCode;
  country_code?: CountryCode;
  address_residential_indicator?: AddressResidentialIndicator;
}